Add support for arbitraryprecision signed decimal number

+++ src/main/java/com/calculator/math/dto/CalculationsDto.java
@@ -1,9 +1,10 @@
 package com.calculator.math.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class CalculationsDto implements Serializable
 {
-    public int a;
-    public int b;
+    public BigDecimal a;
+    public BigDecimal b;
 }
